+++ __test__/useCouponSelected.test.tsx
@@ -1,38 +1,5 @@
-import { activeCouponCodesState, couponSelectedState } from '@store/couponStore';
-import { RecoilRoot, useRecoilValue } from 'recoil';
-import { act, renderHook } from '@testing-library/react';
-import useCouponSelected from '@hooks/coupon/useCouponSelected';
-
-interface renderHookProps {
-  couponSelected: Record<string, boolean>;
-  activeCoupons: string[];
-}
-
-const renderHook_useCouponSelected_activeCoupon = ({
-  couponSelected,
-  activeCoupons,
-}: renderHookProps) => {
-  return renderHook(
-    () => {
-      const { handleToggleCouponCheckbox, couponSelected } = useCouponSelected();
-      const activeCouponCodes = useRecoilValue(activeCouponCodesState);
-
-      return { handleToggleCouponCheckbox, couponSelected, activeCouponCodes };
-    },
-    {
-      wrapper: ({ children }) => (
-        <RecoilRoot
-          initializeState={({ set }) => {
-            set(couponSelectedState, couponSelected);
-            set(activeCouponCodesState, activeCoupons);
-          }}
-        >
-          {children}
-        </RecoilRoot>
-      ),
-    },
-  );
-};
+import { act } from '@testing-library/react';
+import { renderHook_useCouponSelected_activeCoupon } from './renderHook.util';
 
 describe('useCouponSelected hook 테스트', () => {
   it('activeCoupon이 2개인 상황에서 하나의 쿠폰을 더 선택할 수 없다.', () => {
